+++ src/constants.ts
@@ -2,3 +2,6 @@ export const CONFIG_FILES = [".antd4tvrc.cjs"];
 
 export const DEFAULT_OUTPUT_DIR = "node_modules/antd4-theme-vars/themes";
 export const DEFAULT_ANT_PREFIX = "ant";
+
+export const DEFAULT_ANTD_LESS_PATH = "node_modules/antd/dist/antd.less";
+export const DEFAULT_ANTD_LESS_LOOKING_PATH = "node_modules/antd/lib";