+++ src/types.ts
@@ -10,10 +10,16 @@ export type ThemeConfig = {
    * @see https://github.com/ant-design/ant-design/blob/4.x-stable/components/style/themes/default.less
    */
   variables: Record<string, string>;
+
+  /** antd less path, default is `join(process.cwd(), "node_modules/antd/dist/antd.less")`   */
+  antdLessPath?: string;
+
+  /** antd less looking paths, default is `[join(process.cwd(), "node_modules/antd/lib")]` */
+  antdLessLookingPaths?: string[];
 };
 
 export type DefineConfigType = {
-  /** output directory, default is `join(process.cwd(), node_modules/antd4-theme-vars/themes)` */
+  /** output directory, default is `join(process.cwd(), "node_modules/antd4-theme-vars/themes")` */
   outputDir?: string;
 
   /** different theme configs */